feat: refactor logged in header menu to container
