+++ app/components/Layout/Header.tsx
@@ -1,5 +1,4 @@
 import React, {useState, useRef} from 'react';
-import {Form, Dropdown} from 'react-bootstrap';
 import Link from 'next/link';
 import LoginButton from 'components/LoginButton';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
@@ -11,7 +10,7 @@ const HeaderLayout = ({
   isLoggedIn = false,
   isRegistered = false,
   fixed = false,
-  user = null,
+  userProfileDropdown = null,
   children
 }) => {
   let mediaMatch;
@@ -36,12 +35,6 @@ const HeaderLayout = ({
         return;
       return !prev;
     });
-  };
-
-  const submitForm = (e) => {
-    if (e.key === 'Enter' || e.which === 13) {
-      e.currentTarget.submit();
-    }
   };
 
   return (
@@ -90,67 +83,7 @@ const HeaderLayout = ({
               </li>
             ) : null}
             {isLoggedIn ? (
-              <>
-                {isRegistered && (
-                  <>
-                    <li className="d-none d-lg-block">
-                      <Dropdown alignRight>
-                        <Dropdown.Toggle variant="primary">
-                          <FontAwesomeIcon
-                            color="white"
-                            icon={faUser}
-                            size="1x"
-                          />
-                        </Dropdown.Toggle>
-                        <Dropdown.Menu>
-                          <Dropdown.Item as={Link} href="/user/profile">
-                            <a className="dropdown-item text-right">
-                              <div>
-                                <span>{user.firstName}</span>{' '}
-                                <span>{user.lastName}</span>
-                              </div>
-                              <div className="small text-muted text-nowrap">
-                                {user.emailAddress}
-                              </div>
-                            </a>
-                          </Dropdown.Item>
-                          <Dropdown.Item
-                            as={Form}
-                            action="/logout"
-                            method="post"
-                            tabIndex={0}
-                            onKeyPress={submitForm}
-                          >
-                            <button type="submit" className="w-100 text-right">
-                              Logout
-                            </button>
-                          </Dropdown.Item>
-                        </Dropdown.Menu>
-                      </Dropdown>
-                    </li>
-                    <li className="d-lg-none">
-                      <Link href="/user/profile">
-                        <a className="nav-button text-right">
-                          <div>
-                            <span>{user.firstName}</span>{' '}
-                            <span>{user.lastName}</span>
-                          </div>
-                          <div className="small text-nowrap">
-                            {user.emailAddress}
-                          </div>
-                        </a>
-                      </Link>
-                    </li>
-                    <li className="d-lg-none">
-                      <Form action="/logout" method="post">
-                        <button type="submit" className="nav-button text-right">
-                          Logout
-                        </button>
-                      </Form>
-                    </li>
-                  </>
-                )}
-              </>
+              <>{isRegistered && <>{userProfileDropdown}</>}</>
             ) : (
               <>
                 <li>
